feat: 동일한 코드 중복으로 BaseViewModel 구현
