+++ app/src/main/java/com/anshyeon/imagesearchapp/ui/screen/favoriteScreen/FavoriteViewMdoel.kt
@@ -1,21 +1,20 @@
 package com.anshyeon.imagesearchapp.ui.screen.favoriteScreen
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.anshyeon.imagesearchapp.data.model.UnsplashImage
 import com.anshyeon.imagesearchapp.data.repository.ImageRepository
+import com.anshyeon.imagesearchapp.ui.screen.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class FavoriteViewModel @Inject constructor(
     private val imageRepository: ImageRepository
-) : ViewModel() {
+) : BaseViewModel(imageRepository) {
 
     val favoriteUiState: StateFlow<FavoriteUiState> = imageRepository.getFavoriteImages().map {
         FavoriteUiState.Success(it)
@@ -24,16 +23,6 @@ class FavoriteViewModel @Inject constructor(
         started = SharingStarted.WhileSubscribed(5_000),
         initialValue = FavoriteUiState.Loading
     )
-
-    fun toggleFavorite(image: UnsplashImage) {
-        viewModelScope.launch {
-            if (image.isLiked) {
-                imageRepository.deleteImageFromFavorites(image)
-            } else {
-                imageRepository.addImageToFavorites(image)
-            }
-        }
-    }
 }
 
 sealed interface FavoriteUiState {
